pyproject.toml
[project]
name = "vnpt-ai-hackathon-track-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
